+++ blockchain/consensus/fil_cns/src/validation.rs
@@ -1,21 +1,31 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use std::sync::Arc;
+use std::{collections::BTreeMap, sync::Arc};
 
 use cid::Cid;
 use fil_actor_interface::power;
 use fil_actors_runtime_v9::runtime::DomainSeparationTag;
+use filecoin_proofs_api::{post, PublicReplicaInfo, SectorId};
 use forest_beacon::{Beacon, BeaconEntry, BeaconSchedule, IGNORE_DRAND_VAR};
 use forest_blocks::{Block, BlockHeader, Tipset};
 use forest_chain_sync::collect_errs;
 use forest_networks::{ChainConfig, Height};
-use forest_shim::{address::Address, randomness::Randomness, version::NetworkVersion};
-use forest_state_manager::{verify_winning_post, StateManager};
+use forest_shim::{
+    address::Address,
+    randomness::Randomness,
+    sector::{PoStProof, SectorInfo},
+    version::NetworkVersion,
+};
+use forest_state_manager::StateManager;
+use forest_utils::encoding::prover_id_from_u64;
 use futures::stream::FuturesUnordered;
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_encoding::Cbor;
-use fvm_shared::{crypto::signature::ops::verify_bls_sig, TICKET_RANDOMNESS_LOOKBACK};
+use fvm_ipld_encoding::{bytes_32, Cbor};
+use fvm_shared::{
+    commcid::cid_to_replica_commitment_v1, crypto::signature::ops::verify_bls_sig,
+    TICKET_RANDOMNESS_LOOKBACK,
+};
 use nonempty::NonEmpty;
 
 use crate::{metrics, FilecoinConsensusError};
@@ -393,4 +403,65 @@ fn verify_winning_post_proof<DB: Blockstore + Clone + Send + Sync + 'static>(
         id,
     )
     .map_err(|e| FilecoinConsensusError::WinningPoStValidation(e.to_string()))
+}
+
+fn to_fil_public_replica_infos(
+    src: &[SectorInfo],
+    typ: ProofType,
+) -> Result<BTreeMap<SectorId, PublicReplicaInfo>, String> {
+    let replicas = src
+        .iter()
+        .map::<Result<(SectorId, PublicReplicaInfo), String>, _>(|sector_info: &SectorInfo| {
+            let commr = cid_to_replica_commitment_v1(&sector_info.sealed_cid)?;
+            let proof = match typ {
+                ProofType::Winning => sector_info.proof.registered_winning_post_proof()?,
+                // ProofType::Window => sector_info.proof.registered_window_post_proof()?,
+            };
+            let replica = PublicReplicaInfo::new(proof.try_into()?, commr);
+            Ok((SectorId::from(sector_info.sector_number), replica))
+        })
+        .collect::<Result<BTreeMap<SectorId, PublicReplicaInfo>, _>>()?;
+    Ok(replicas)
+}
+
+/// Functionality for verification of seal, winning PoSt and window PoSt proofs.
+/// Proof verification will be full validation by default.
+
+/// Verifies winning proof of spacetime. These proofs are generated by the
+/// miners that are elected to mine a new block to verify a sector. A failed
+/// winning proof leads to a miner being slashed.
+
+/// PoSt proof variants.
+enum ProofType {
+    Winning,
+    // Window,
+}
+
+pub fn verify_winning_post(
+    mut rand: Randomness,
+    proofs: &[PoStProof],
+    challenge_sectors: &[SectorInfo],
+    prover: u64,
+) -> Result<(), anyhow::Error> {
+    // Necessary to be valid bls12 381 element.
+    rand.0[31] &= 0x3f;
+
+    // Convert sector info into public replica
+    let replicas = to_fil_public_replica_infos(challenge_sectors, ProofType::Winning)
+        .map_err(|e| anyhow::anyhow!("{}", e))?;
+
+    // Convert PoSt proofs into proofs-api format
+    let proof_bytes = proofs.iter().fold(Vec::new(), |mut proof, p| {
+        proof.extend_from_slice(&p.proof_bytes);
+        proof
+    });
+
+    // Generate prover bytes from ID
+    let prover_id = prover_id_from_u64(prover);
+
+    // Verify Proof
+    if !post::verify_winning_post(&bytes_32(&rand.0), &proof_bytes, &replicas, prover_id)? {
+        anyhow::bail!("Winning post was invalid")
+    }
+    Ok(())
 }
