+++ utils/forest_utils/Cargo.toml
@@ -15,6 +15,8 @@ blake2b_simd.workspace = true
 cid.workspace = true
 const_format = "0.2.26"
 digest.workspace = true
+filecoin-proofs-api.workspace = true
+forest_shim.workspace = true
 futures.workspace = true
 fvm_ipld_blockstore.workspace = true
 fvm_ipld_encoding.workspace = true
